refactor: EventDetail, Tag 패키지 분리
