+++ src/test/java/com/example/donttouchme/excel/integration/ExcelQueryServiceTest.java
@@ -1,13 +1,15 @@
 package com.example.donttouchme.excel.integration;
 
 import com.example.donttouchme.event.domain.*;
-import com.example.donttouchme.event.repository.EventDetailRepository;
+import com.example.donttouchme.eventdetail.repository.EventDetailRepository;
 import com.example.donttouchme.event.repository.EventRepository;
-import com.example.donttouchme.event.repository.TagRepository;
+import com.example.donttouchme.tag.repository.TagRepository;
+import com.example.donttouchme.eventdetail.domain.EventDetail;
 import com.example.donttouchme.excel.service.ExcelQueryService;
 import com.example.donttouchme.member.domain.Member;
 import com.example.donttouchme.member.repository.MemberRepository;
 import com.example.donttouchme.support.IntegrationTestSupport;
+import com.example.donttouchme.tag.domain.Tag;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
